refactor(app): lift the empty-thread state out of ChatPanel
The 75-line JSX blob behind `emptyState` — per-session empty thread, the
"you picked an actor, now type" draft hint, and the local-agent welcome
card — moves to `renderChatEmptyState`. Its props are exactly the ten
values the memo already listed as dependencies.

It is deliberately a plain function and not a component. `MessageList`
branches on `emptyState === null` ("render nothing at all") versus
non-null ("center an empty-state block"), so the `return null` paths here
are load-bearing: a component would always hand back a non-null element
and silently turn every one of them into a centered empty block. The call
also stays inside `React.useMemo`, because `MessageList` takes the result
as a prop and a fresh node each render would defeat its memoization. Both
constraints are written down in the new file.

ChatPanel 1,578 -> 1,504 lines.
